requirements.txt
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
slack-sdk==3.27.1
